tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-timing-function_multiple_keyframes.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-timing-function, multiple keyframes</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animation-timing-function_tag">
   <meta name="flags" content="dom">
   <meta name="assert" content="For a keyframed animation, the
   'animation-timing-function' applies between keyframes, not over the entire
   animation">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move {
         0% {
            margin-left: 0px;
         }
         50% {
            margin-left: 100px;
         }
         100% {
            margin-left: 200px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         animation-fill-mode: forwards;
         animation-duration: 1s;
         animation-timing-function: step-end;
      }
      .test {
         animation-name: move;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var style = getComputedStyle(div, null);
         var passed = false;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         setTimeout(function() {
            passed = (style.marginLeft == '100px');
         }, 750);
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            passed = (passed && style.marginLeft == '200px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should jump rightwards in two steps.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
